requirements.txt
fastapi
uvicorn
httpx
python-dotenv
coinbase-advanced-py
